Add a Definition schema object that contains field metadata
This new Definition class, along with a helper define function,
allows to define schema fields with multiple constraints. It also
introduces a concept of default value which can be assigned by
either providing the default value primitive or a factory function.

A _sentinel object is also introduced to provide a value for absent
fields, so that the null (None) can also be a valid field value.